requirements.txt
openai
supabase-py
python-dotenv
requests
tqdm
colorama 